fix some validation error

+++ src/app/modules/blog/blog.controller.ts
@@ -12,7 +12,7 @@ const createBlog = catchAsync(async (req, res) => {
   const token = req.headers.authorization;
   if (!token) {
     throw new AppError(status.UNAUTHORIZED, 'unauthorized');
-  }
+  } 
   const decoded = jwt.verify(
     token,
     config.jwt_access_secret as string,
@@ -38,9 +38,22 @@ const createBlog = catchAsync(async (req, res) => {
 const updateBlog = catchAsync(async (req, res) => {
   const id = req.params.id;
   const blog = req?.body;
+  const token = req.headers.authorization;
+  if (!token) {
+    throw new AppError(status.UNAUTHORIZED, 'unauthorized');
+  } 
   const isBlogExit = await Blog.findById(id);
   if (!isBlogExit) {
     throw new AppError(status.NOT_FOUND, 'Blog not found');
+  }
+  const decoded = jwt.verify(
+    token,
+    config.jwt_access_secret as string,
+  ) as JwtPayload;
+  const { userId } = decoded;
+  const author = isBlogExit?.author?.toString();
+  if(userId !== author){
+    throw new AppError(status.FORBIDDEN, 'You are not allowed to update this blog');
   }
   const result = await blogServices.updateBlogIntoDB(id, blog);
   res.status(200).json({
@@ -52,9 +65,22 @@ const updateBlog = catchAsync(async (req, res) => {
 
 const deleteBlog = catchAsync(async (req, res) => {
   const id = req.params.id;
+  const token = req.headers.authorization;
+  if (!token) {
+    throw new AppError(status.UNAUTHORIZED, 'unauthorized');
+  } 
   const isBlogExit = await Blog.findById(id);
   if (!isBlogExit) {
     throw new AppError(status.NOT_FOUND, 'Blog not found');
+  }
+  const decoded = jwt.verify(
+    token,
+    config.jwt_access_secret as string,
+  ) as JwtPayload;
+  const { userId } = decoded;
+  const author = isBlogExit?.author?.toString();
+  if(userId !== author){
+    throw new AppError(status.FORBIDDEN, 'You are not allowed to delete this blog');
   }
    await blogServices.deleteBlogFromDB(id);
   res.status(200).json({
